bash with descriptions
